requirements.txt
atproto
beautifulsoup4
google-generativeai
Pillow
requests
